Small fixes (Lost time because 0.5 doesn't work but 0,5 does...)

+++ VivesShop.Ui.Mvc/Models/CartProduct.cs
@@ -9,7 +9,7 @@ namespace VivesShop.Ui.Mvc.Models
         [DisplayName("Product name")]
         public required string Name { get; set; }
         [DisplayName("Price")]
-        public required float Price { get; set; }
+        public required decimal Price { get; set; }
 
     }
 }
